[fix] Handle None when print a table
When printing a table (e.g.: `CodeChecker version`) it is possible
that a value is None (e.g.: a package is built from a release zip file
which will not contain git information). In this case when printing the
table the following exception will be raised:
```
unsupported format string passed to NoneType.__format__
```
In this patch we will convert None values to empty string to handle this
use case too.